dags/immo-eliza/dashboard/requirements.txt
azure-core==1.26.1
azure-storage-blob==12.14.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-bootstrap-templates==1.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
plotly==5.11.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tenacity==8.1.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
